requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
pandas==2.1.3
plotly==5.17.0
pyyaml==6.0.1
python-dateutil==2.8.2
schedule==1.2.0
yagmail==0.15.293
python-dotenv==1.0.0
selenium==4.15.2
openpyxl==3.1.2
google-generativeai==0.3.2
fpdf==1.7.2
openai>=1.0.0